:recycle: Separate file for RagnarokCoreError
